refactor: mv state utils to utils
